+++ HISTORY.rst
@@ -26,6 +26,7 @@ latest (unreleased)
 * Empty company on products, all products should now be shared (MIGO-328)
 
 * Activate migrated batch pickings
+* Cancel french draft invoices (MIGO-334)
 
 **Features and Improvements**